requirements.txt
opencv-python~=4.5.2.54
Pillow~=8.3.0
Flask~=2.0.1
WTForms~=2.3.3
Werkzeug~=2.0.1
scikit-learn~=0.24.2
colormap~=1.0.3